+++ CHANGELOG.md
@@ -20,15 +20,15 @@ _Release date: 1 Sep 2018_
     - [x] SHORTCUT: add 'Ctrl+Q' to quit the app 
     - [x] BUG: fix left alt value error
     - [ ] UI: make it run in the system tray
-    - [ ] UI: find a suitable icon
+    - [x] UI: find a suitable icon
     - [x] IMPROVEMENT: self.settings should be updated on the go
     - [x] FEEDBACK: auto-sort when gestures are updated
-    - [ ] REMOVE BUTTON: disable and enable when a gesture in the Gesture column is selected (2) should not ask what gesture to delete
+    - [x] REMOVE BUTTON: disable and enable when a gesture in the Gesture column is selected (2) should not ask what gesture to delete
     - [x] REMOVE BUTTON: self-destruct when no existing gestures is remove
     - [x] UPDATE: not reflecting after updating either gesture or meaning
     - [x] GESTURESTABLE: should retain the previous value of the cell editing
     - [x] GESTURESTABLE: get selectedData via using the arrow keys
-    - [ ] ADD BUTTON: you can still add existing gestures
+    - [x] ADD BUTTON: you can still add existing gestures 
 
 
 **Patch 1.1.1**